By the numbers complete
