+++ src/frontend.js
@@ -12,8 +12,5 @@ import './features/core-group-report-overlapping-callout';
 // Open accordion first items by default.
 import './features/accordion';
 
-// Render animations.
-import './features/animations';
-
-// By the numbers masonry.
-import './features/masonry';
+// By the numbers masonry and animations.
+import './features/by-the-numbers';
